ci: split the five files over the 500-LoC cap
All five are this branch's own growth, so no allow-list entries were added —
the cap's allow-list wants a rationale AND a tracking issue, and "I just wrote
this" is neither.

  layers/ngram_embed.rs         864 -> 109  + ids/table/embed/tests
  weight_loader/longcat/mod.rs  564 -> 436  + ffn.rs
  spark-runtime/weights.rs      536 -> 420  + weights/name_utils.rs
  model/impl_a3.rs              518 -> 472  + impl_a3_embed.rs
  layers/moe/forward_prefill.rs 506 -> 491  + forward_prefill_router.rs

Split along real seams rather than by line count:

  ngram_embed/ids.rs     the hash id core — pure token-id arithmetic, no
                         device, no weights, which is exactly why it can be
                         tested bit-exactly against the Python reference
  ngram_embed/table.rs   the per-table storage enum (BF16 / FP8 / NVMe-cached)
  ngram_embed/embed.rs   the fused GPU embedding itself
  longcat/ffn.rs         the two per-sublayer FFN builders — the dense SwiGLU
                         every sublayer has, and the shortcut MoE sublayer 0
                         computes for sublayer 1 to add
  weights/name_utils.rs  tensor-NAME parsing: numeric trailing-index sort, the
                         n-gram predicate, expert-index extraction
  impl_a3_embed.rs       single-token embedding, including the guard that
                         REFUSES a bare token id on an n-gram model
  forward_prefill_router LongCat's softmax+bias router with the zero-expert
                         fold, batched-prefill side

Mechanical fallout worth knowing for the next split of this kind: moving a
`#[cfg(test)] mod tests { use super::*; }` into its own FILE silently changes
what `super` means — it becomes the tests module itself — so the imports must
be re-anchored at the real parent, and the now-redundant inner `mod tests`
wrapper trips clippy's `module_inception`. Fields the old single file shared
freely became private across the new boundary and are now `pub(super)`, which
keeps them invisible outside the module.

Verified beyond compiling: workspace clippy --all-targets is clean (that also
picked up an unrelated empty-format-string in a microtest example), fmt is
clean under --check, 648 spark-model + 276 spark-runtime tests pass, and the
full model still serves byte-identical output to the pre-split run
("C. Paris").
