+++ crates/spark-model/src/layers/moe/mod.rs
@@ -479,6 +479,7 @@ mod forward_prefill_bf16;
 mod forward_prefill_fp8;
 mod forward_prefill_phase;
 mod forward_prefill_routed;
+mod forward_prefill_router;
 mod forward_token_major;
 mod helpers_a;
 mod helpers_b;
