+++ crates/spark-model/src/layers/ngram_embed.rs
@@ -96,769 +96,13 @@ impl NgramDims {
 /// span [t-n, t] crosses a document boundary (EOS-inclusive segment end),
 /// in which case 0. Mirrors the reference implementation exactly, including
 /// the quirk that a segment shorter than n contributes nothing.
-fn shift_right_ignore_eos(ctx: &[u32], n: usize, eos: u32) -> Vec<u32> {
-    let len = ctx.len();
-    let mut out = vec![0u32; len];
-    let mut prev = 0usize;
-    for (pos, &tok) in ctx.iter().enumerate() {
-        if tok == eos {
-            let end = pos + 1;
-            if end - prev > n {
-                out[prev + n..end].copy_from_slice(&ctx[prev..end - n]);
-            }
-            prev = end;
-        }
-    }
-    if prev < len && len - prev > n {
-        out[prev + n..len].copy_from_slice(&ctx[prev..len - n]);
-    }
-    out
-}
+mod embed;
+mod ids;
+mod table;
 
-/// Compute the row ids for EVERY table over `ctx` (the n-1 cached context
-/// tokens followed by the new tokens). Returns `num_tables` vectors of
-/// `ctx.len()` ids each, table-major in reference index order
-/// (`(ngram-2)*K + split`); callers slice the last `seq_len` entries.
-pub fn ngram_ids(dims: &NgramDims, ctx: &[u32]) -> Vec<Vec<u64>> {
-    let mut out = Vec::with_capacity(dims.num_tables());
-    // shift_d computed once per d, shared across splits (reference computes
-    // shifted_ids once per n-gram size; d ranges 1..=N-1).
-    let shifts: Vec<Vec<u32>> = (1..dims.neighbor_num)
-        .map(|d| shift_right_ignore_eos(ctx, d, dims.eos_token_id))
-        .collect();
-    for ngram in 2..=dims.neighbor_num {
-        for split in 0..dims.split_num {
-            let index = (ngram - 2) * dims.split_num + split;
-            let t = dims.table_rows(index);
-            let mods = dims.vocab_mods(ngram, split);
-            let ids = ctx
-                .iter()
-                .enumerate()
-                .map(|(pos, &x)| {
-                    let mut acc = x as u64;
-                    for (d, &m) in mods.iter().enumerate() {
-                        acc += shifts[d][pos] as u64 * m;
-                    }
-                    acc % t
-                })
-                .collect();
-            out.push(ids);
-        }
-    }
-    out
-}
-
-// ── GPU module ───────────────────────────────────────────────────────────
-
-use anyhow::{Context, Result};
-use spark_runtime::gpu::{DevicePtr, GpuBackend, KernelHandle};
-
-use crate::weight_map::{DenseWeight, Fp8DenseWeight};
-
-/// One n-gram lookup table on device: BF16 as shipped, or FP8-quantized at
-/// load (per-row E4M3 + f32 scale — halves the ~63 GB table footprint;
-/// embeddings tolerate this well and the gather dequantizes on read).
-pub enum NgramTable {
-    Bf16(DenseWeight),
-    Fp8(Fp8DenseWeight),
-    /// NVMe-backed: only a bounded set of ROWS is resident, in a pinned
-    /// GPU-addressable arena. The host resolves `row_id -> slot` (the ids are
-    /// a pure function of token ids, so this is host-side anyway) and the
-    /// SAME gather kernels then read the arena by slot index — no kernel
-    /// change, no `cuMemcpyHtoD` on the fault path.
-    ///
-    /// This is what makes a 51 B-parameter embedding table serveable on a
-    /// 121 GB box: the tables are the model's largest tensors and its least
-    /// bandwidth-hungry (12 rows ~ 3 KB per token), so demoting them buys
-    /// back tens of GB for KV.
-    Cached(Box<spark_storage::NgramRowCache>),
-}
-
-impl NgramTable {
-    /// Quantize a BF16 table to FP8 on the GPU (per-row E4M3 absmax +
-    /// f32 scale via `quantize_bf16_to_fp8`) — the quantize-on-load
-    /// lever. The caller frees the BF16 source afterwards; tables are
-    /// loaded one at a time so peak overhead is a single table.
-    pub fn quantize_bf16(
-        w: &DenseWeight,
-        rows: usize,
-        dim: usize,
-        gpu: &dyn GpuBackend,
-        stream: u64,
-    ) -> Result<Self> {
-        let k = gpu.kernel("gemv_fp8w", "quantize_bf16_to_fp8")?;
-        Ok(Self::Fp8(crate::weight_map::quantize_to_fp8(
-            w, rows, dim, gpu, k, stream,
-        )?))
-    }
-}
-
-/// GPU n-gram embedding: base word embedding fused with the K*(N-1)
-/// hashed-table lookups, composed entirely from existing kernels
-/// (`batched_embed` gathers + `dense_gemm_bf16_pipelined` projections +
-/// `bf16_scaled_add` accumulation). A fused single-kernel version is a
-/// later optimization; per-token cost here is 1+T gathers, T tiny GEMMs
-/// and T+1 scaled adds (T = 12 for LongCat-Lite) — negligible next to a
-/// transformer layer.
-pub struct NgramEmbedding {
-    pub dims: NgramDims,
-    /// Base word embedding `[vocab, hidden]` BF16.
-    pub word: DenseWeight,
-    /// The K*(N-1) lookup tables, index order `(ngram-2)*K + split`,
-    /// each `[table_rows(i), table_dim]` — BF16 or FP8-quantized.
-    pub tables: Vec<NgramTable>,
-    /// Per-table projections `[hidden, table_dim]` BF16 (nn.Linear layout).
-    pub projs: Vec<DenseWeight>,
-
-    batched_embed_k: KernelHandle,
-    batched_embed_fp8_k: KernelHandle,
-    gemm_k: KernelHandle,
-    scaled_add_k: KernelHandle,
-
-    /// Device staging: ids `[max_tokens]` u32 (reused per table),
-    /// gathered rows `[max_tokens, table_dim]` BF16, projected rows
-    /// `[max_tokens, hidden]` BF16.
-    ids_dev: DevicePtr,
-    gather_buf: DevicePtr,
-    proj_buf: DevicePtr,
-    max_tokens: usize,
-}
-
-impl NgramEmbedding {
-    pub fn new(
-        dims: NgramDims,
-        word: DenseWeight,
-        tables: Vec<NgramTable>,
-        projs: Vec<DenseWeight>,
-        max_tokens: usize,
-        gpu: &dyn GpuBackend,
-    ) -> Result<Self> {
-        anyhow::ensure!(tables.len() == dims.num_tables(), "table count");
-        anyhow::ensure!(projs.len() == dims.num_tables(), "proj count");
-        let td = dims.table_dim();
-        Ok(Self {
-            dims,
-            word,
-            tables,
-            projs,
-            batched_embed_k: gpu.kernel("embed_from_argmax", "batched_embed")?,
-            batched_embed_fp8_k: gpu.kernel("embed_from_argmax", "batched_embed_fp8")?,
-            gemm_k: gpu.kernel("gemm", "dense_gemm_bf16_pipelined")?,
-            scaled_add_k: gpu.kernel("residual_add", "bf16_scaled_add")?,
-            ids_dev: gpu.alloc(max_tokens * 4)?,
-            gather_buf: gpu.alloc(max_tokens * td * 2)?,
-            proj_buf: gpu.alloc(max_tokens * dims.hidden_size * 2)?,
-            max_tokens,
-        })
-    }
-
-    /// Fused embedding for the LAST `seq_len` tokens of `ctx` (`ctx` =
-    /// up to n-1 cached context tokens followed by the new tokens —
-    /// exactly the reference NgramCache contract). Writes
-    /// `[seq_len, hidden]` BF16 to `out`.
-    pub fn embed(
-        &mut self,
-        ctx_tokens: &[u32],
-        seq_len: usize,
-        out: DevicePtr,
-        gpu: &dyn GpuBackend,
-        stream: u64,
-    ) -> Result<()> {
-        use crate::layers::ops;
-        anyhow::ensure!(
-            seq_len <= self.max_tokens,
-            "ngram embed: seq_len over staging"
-        );
-        anyhow::ensure!(seq_len <= ctx_tokens.len(), "ngram embed: seq_len over ctx");
-        let h = self.dims.hidden_size;
-        let td = self.dims.table_dim();
-        let inv_scale = 1.0f32 / (1 + self.dims.num_tables()) as f32;
-
-        // out = 0; each contribution lands via scaled_add(out += src/13),
-        // matching the reference's final 1/(1+T) over base + all tables.
-        gpu.memset(out, 0, seq_len * h * 2)?;
-
-        // Base word rows for the NEW tokens only.
-        let new_tokens = &ctx_tokens[ctx_tokens.len() - seq_len..];
-        let ids_bytes: Vec<u8> = new_tokens.iter().flat_map(|t| t.to_le_bytes()).collect();
-        gpu.copy_h2d_async(&ids_bytes, self.ids_dev, stream)?;
-        ops::batched_embed(
-            gpu,
-            self.batched_embed_k,
-            self.ids_dev,
-            self.word.weight,
-            self.proj_buf,
-            seq_len as u32,
-            h as u32,
-            stream,
-        )?;
-        ops::scaled_add(
-            gpu,
-            self.scaled_add_k,
-            out,
-            self.proj_buf,
-            inv_scale,
-            (seq_len * h) as u32,
-            stream,
-        )?;
-
-        // Host-side hash ids over the full ctx, one table at a time.
-        let all_ids = ngram_ids(&self.dims, ctx_tokens);
-        for (index, ids) in all_ids.iter().enumerate() {
-            let tail = &ids[ids.len() - seq_len..];
-            let id_bytes: Vec<u8> = tail
-                .iter()
-                .map(|&v| u32::try_from(v).context("ngram id exceeds u32"))
-                .collect::<Result<Vec<u32>>>()?
-                .iter()
-                .flat_map(|v| v.to_le_bytes())
-                .collect();
-            gpu.copy_h2d_async(&id_bytes, self.ids_dev, stream)?;
-            // NVMe-backed table: fault the rows in and REPLACE the ids with
-            // their slot indices, then gather from the arena exactly as if it
-            // were a small resident table.
-            if let NgramTable::Cached(cache) = &mut self.tables[index] {
-                let mut slots = Vec::with_capacity(seq_len);
-                cache.resolve(tail, &mut slots)?;
-                let slot_bytes: Vec<u8> = slots.iter().flat_map(|v| v.to_le_bytes()).collect();
-                gpu.copy_h2d_async(&slot_bytes, self.ids_dev, stream)?;
-                let table = DevicePtr(cache.table_dev_va()?);
-                match cache.scale_dev_va()? {
-                    Some(sc) => ops::batched_embed_fp8(
-                        gpu,
-                        self.batched_embed_fp8_k,
-                        self.ids_dev,
-                        table,
-                        DevicePtr(sc),
-                        self.gather_buf,
-                        seq_len as u32,
-                        td as u32,
-                        stream,
-                    )?,
-                    None => ops::batched_embed(
-                        gpu,
-                        self.batched_embed_k,
-                        self.ids_dev,
-                        table,
-                        self.gather_buf,
-                        seq_len as u32,
-                        td as u32,
-                        stream,
-                    )?,
-                }
-                cache.end_batch();
-                ops::dense_gemm_bf16_pipelined(
-                    gpu,
-                    self.gemm_k,
-                    self.gather_buf,
-                    &self.projs[index],
-                    self.proj_buf,
-                    seq_len as u32,
-                    h as u32,
-                    td as u32,
-                    stream,
-                )?;
-                ops::scaled_add(
-                    gpu,
-                    self.scaled_add_k,
-                    out,
-                    self.proj_buf,
-                    inv_scale,
-                    (seq_len * h) as u32,
-                    stream,
-                )?;
-                continue;
-            }
-            match &self.tables[index] {
-                NgramTable::Bf16(w) => ops::batched_embed(
-                    gpu,
-                    self.batched_embed_k,
-                    self.ids_dev,
-                    w.weight,
-                    self.gather_buf,
-                    seq_len as u32,
-                    td as u32,
-                    stream,
-                )?,
-                NgramTable::Fp8(w) => ops::batched_embed_fp8(
-                    gpu,
-                    self.batched_embed_fp8_k,
-                    self.ids_dev,
-                    w.weight,
-                    w.row_scale,
-                    self.gather_buf,
-                    seq_len as u32,
-                    td as u32,
-                    stream,
-                )?,
-                NgramTable::Cached(_) => unreachable!("resolved above"),
-            }
-            ops::dense_gemm_bf16_pipelined(
-                gpu,
-                self.gemm_k,
-                self.gather_buf,
-                &self.projs[index],
-                self.proj_buf,
-                seq_len as u32,
-                h as u32,
-                td as u32,
-                stream,
-            )?;
-            ops::scaled_add(
-                gpu,
-                self.scaled_add_k,
-                out,
-                self.proj_buf,
-                inv_scale,
-                (seq_len * h) as u32,
-                stream,
-            )?;
-        }
-        Ok(())
-    }
-}
+pub use embed::NgramEmbedding;
+pub use ids::ngram_ids;
+pub use table::NgramTable;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[derive(serde::Deserialize)]
-    struct Fixture {
-        name: String,
-        vocab_size: u64,
-        hidden_size: usize,
-        ngram_vocab_size_ratio: u64,
-        emb_neighbor_num: usize,
-        emb_split_num: usize,
-        eos_token_id: u32,
-        tokens: Vec<u32>,
-        expected_ids: std::collections::HashMap<String, Vec<u64>>,
-    }
-
-    #[test]
-    fn ids_match_python_reference() {
-        let path = concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/../../bench/ngram_ref/ngram_id_fixtures.json"
-        );
-        let fixtures: Vec<Fixture> = serde_json::from_str(
-            &std::fs::read_to_string(path)
-                .expect("run bench/ngram_ref/make_fixtures.py to generate fixtures"),
-        )
-        .unwrap();
-        assert!(!fixtures.is_empty());
-        for f in fixtures {
-            let dims = NgramDims {
-                vocab_size: f.vocab_size,
-                ratio: f.ngram_vocab_size_ratio,
-                neighbor_num: f.emb_neighbor_num,
-                split_num: f.emb_split_num,
-                eos_token_id: f.eos_token_id,
-                hidden_size: f.hidden_size,
-            };
-            let got = ngram_ids(&dims, &f.tokens);
-            assert_eq!(got.len(), dims.num_tables(), "{}", f.name);
-            for (index, ids) in got.iter().enumerate() {
-                let want = &f.expected_ids[&index.to_string()];
-                assert_eq!(ids, want, "{} table {index}", f.name);
-            }
-        }
-    }
-
-    /// GPU parity vs the real-checkpoint golden vectors, on COMPACTED
-    /// tables (only the rows the fixture touches — ~20 MB instead of
-    /// 63 GB). Data dir comes from `ATLAS_NGRAM_TEST_DATA`, generated by
-    /// `bench/ngram_ref/make_gpu_testdata.py`. GPU test: `#[ignore]` per
-    /// repo convention (CI is CPU-only); run with
-    ///   ATLAS_NGRAM_TEST_DATA=/tank/atlas-testdata/ngram_longcat_lite \
-    ///   cargo test -p spark-model --release ngram_gpu -- --ignored
-    #[test]
-    #[ignore]
-    fn ngram_gpu_matches_golden() {
-        let dir = std::env::var("ATLAS_NGRAM_TEST_DATA")
-            .expect("set ATLAS_NGRAM_TEST_DATA (see make_gpu_testdata.py)");
-        let meta: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(format!("{dir}/meta.json")).unwrap())
-                .unwrap();
-        let gpu =
-            spark_runtime::cuda_backend::AtlasCudaBackend::new(0, &atlas_kernels::ptx_modules())
-                .expect("CUDA backend");
-        let g: &dyn GpuBackend = &gpu;
-
-        let toks: Vec<u32> = meta["tokens"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|v| v.as_u64().unwrap() as u32)
-            .collect();
-        let hidden = meta["hidden_size"].as_u64().unwrap() as usize;
-        let dims = NgramDims {
-            vocab_size: meta["vocab_size"].as_u64().unwrap(),
-            ratio: meta["ngram_vocab_size_ratio"].as_u64().unwrap(),
-            neighbor_num: meta["emb_neighbor_num"].as_u64().unwrap() as usize,
-            split_num: meta["emb_split_num"].as_u64().unwrap() as usize,
-            eos_token_id: meta["eos_token_id"].as_u64().unwrap() as u32,
-            hidden_size: hidden,
-        };
-
-        let upload = |bytes: &[u8]| -> DevicePtr {
-            let p = g.alloc(bytes.len()).unwrap();
-            g.copy_h2d_async(bytes, p, g.default_stream()).unwrap();
-            p
-        };
-        let load = |path: String| std::fs::read(path).unwrap();
-
-        // Compacted word table + remapped base token ids.
-        let word_ids: Vec<u64> = meta["word_ids"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|v| v.as_u64().unwrap())
-            .collect();
-        let word = DenseWeight {
-            weight: upload(&load(format!("{dir}/word_rows.bin"))),
-        };
-        let remap = |ids: &[u64], all: &[u64]| -> Vec<u32> {
-            ids.iter()
-                .map(|id| all.binary_search(id).unwrap() as u32)
-                .collect()
-        };
-        let compact_toks = remap(
-            &toks.iter().map(|&t| t as u64).collect::<Vec<_>>(),
-            &word_ids,
-        );
-
-        // Compacted per-table rows + full projections; the test computes
-        // the SAME hash ids the module will and remaps them to the compact
-        // row order — so the id math is exercised end to end.
-        let host_ids = ngram_ids(&dims, &toks);
-        let mut tables = Vec::new();
-        let mut projs = Vec::new();
-        let mut compact_ids = Vec::new();
-        let mut table_rows_n = Vec::new();
-        for t in meta["tables"].as_array().unwrap() {
-            let index = t["index"].as_u64().unwrap() as usize;
-            assert_eq!(index, tables.len());
-            let ids: Vec<u64> = t["ids"]
-                .as_array()
-                .unwrap()
-                .iter()
-                .map(|v| v.as_u64().unwrap())
-                .collect();
-            tables.push(NgramTable::Bf16(DenseWeight {
-                weight: upload(&load(format!("{dir}/table{index}_rows.bin"))),
-            }));
-            table_rows_n.push(ids.len());
-            projs.push(DenseWeight {
-                weight: upload(&load(format!("{dir}/proj{index}.bin"))),
-            });
-            compact_ids.push(remap(&host_ids[index], &ids));
-        }
-
-        // Build the module on the compacted tables, then OVERRIDE the id
-        // path by embedding through a shim: we can't inject compact ids
-        // into embed() (it recomputes real ids), so drive the same op
-        // sequence manually here — this validates the GPU op composition
-        // and the id math jointly (ids were remapped from OUR ngram_ids).
-        let ng = NgramEmbedding::new(dims, word, tables, projs, 32, g).unwrap();
-        let stream = g.default_stream();
-        let m = toks.len();
-        let out = g.alloc(m * hidden * 2).unwrap();
-        let inv = 1.0f32 / (1 + ng.dims.num_tables()) as f32;
-        let td = ng.dims.table_dim();
-        let put_ids = |ids: &[u32]| {
-            let b: Vec<u8> = ids.iter().flat_map(|v| v.to_le_bytes()).collect();
-            g.copy_h2d_async(&b, ng.ids_dev, stream).unwrap();
-        };
-        let run = |tables: &[NgramTable]| -> Vec<u8> {
-            g.memset(out, 0, m * hidden * 2).unwrap();
-            put_ids(&compact_toks);
-            crate::layers::ops::batched_embed(
-                g,
-                ng.batched_embed_k,
-                ng.ids_dev,
-                ng.word.weight,
-                ng.proj_buf,
-                m as u32,
-                hidden as u32,
-                stream,
-            )
-            .unwrap();
-            crate::layers::ops::scaled_add(
-                g,
-                ng.scaled_add_k,
-                out,
-                ng.proj_buf,
-                inv,
-                (m * hidden) as u32,
-                stream,
-            )
-            .unwrap();
-            for index in 0..ng.dims.num_tables() {
-                put_ids(&compact_ids[index]);
-                match &tables[index] {
-                    NgramTable::Bf16(w) => crate::layers::ops::batched_embed(
-                        g,
-                        ng.batched_embed_k,
-                        ng.ids_dev,
-                        w.weight,
-                        ng.gather_buf,
-                        m as u32,
-                        td as u32,
-                        stream,
-                    )
-                    .unwrap(),
-                    NgramTable::Fp8(w) => crate::layers::ops::batched_embed_fp8(
-                        g,
-                        ng.batched_embed_fp8_k,
-                        ng.ids_dev,
-                        w.weight,
-                        w.row_scale,
-                        ng.gather_buf,
-                        m as u32,
-                        td as u32,
-                        stream,
-                    )
-                    .unwrap(),
-                    // The NVMe-backed variant has its own parity test
-                    // (`cached_table_matches_resident_table`); this manual
-                    // op-sequence harness drives resident tables only.
-                    NgramTable::Cached(_) => unreachable!("test builds resident tables"),
-                }
-                crate::layers::ops::dense_gemm_bf16_pipelined(
-                    g,
-                    ng.gemm_k,
-                    ng.gather_buf,
-                    &ng.projs[index],
-                    ng.proj_buf,
-                    m as u32,
-                    hidden as u32,
-                    td as u32,
-                    stream,
-                )
-                .unwrap();
-                crate::layers::ops::scaled_add(
-                    g,
-                    ng.scaled_add_k,
-                    out,
-                    ng.proj_buf,
-                    inv,
-                    (m * hidden) as u32,
-                    stream,
-                )
-                .unwrap();
-            }
-            g.synchronize(stream).unwrap();
-            let mut got = vec![0u8; m * hidden * 2];
-            g.copy_d2h(out, &mut got).unwrap();
-            got
-        };
-
-        // Compare vs golden f32. Criterion: per-row relative Frobenius
-        // error — elementwise max-rel is the wrong instrument here: the
-        // GPU path accumulates 13 BF16-rounded adds while the golden
-        // accumulates in f32, and simulating that exact chain over the
-        // reference itself (bench/ngram_ref/sim_bf16_chain.py) gives
-        // elementwise max_rel 0.17 with per-row Frobenius only 0.0067.
-        // The 2% gate keeps ~3x headroom over measured rounding noise
-        // while a single mis-mapped table would register ~27% (1/13 of
-        // the row's energy replaced).
-        let golden = load(format!("{dir}/golden_f32.bin"));
-        let worst_row = |got_bf16: &[u8]| -> f32 {
-            let mut worst = 0f32;
-            for r in 0..m {
-                let mut err2 = 0f64;
-                let mut ref2 = 0f64;
-                for c in 0..hidden {
-                    let i = r * hidden + c;
-                    let bits = u16::from_le_bytes([got_bf16[i * 2], got_bf16[i * 2 + 1]]);
-                    let got = f32::from_bits((bits as u32) << 16) as f64;
-                    let want = f32::from_le_bytes([
-                        golden[i * 4],
-                        golden[i * 4 + 1],
-                        golden[i * 4 + 2],
-                        golden[i * 4 + 3],
-                    ]) as f64;
-                    err2 += (got - want) * (got - want);
-                    ref2 += want * want;
-                }
-                worst = worst.max((err2 / ref2).sqrt() as f32);
-            }
-            worst
-        };
-
-        let bf16_worst = worst_row(&run(&ng.tables));
-        assert!(
-            bf16_worst < 0.02,
-            "BF16 fused embedding diverges from golden: worst row Frobenius \
-             rel = {bf16_worst}"
-        );
-        println!("ngram GPU parity (BF16 tables): worst row Frobenius rel = {bf16_worst:.4}");
-
-        // FP8 leg: quantize the compacted tables on the GPU (the
-        // quantize-on-load path) and rerun. E4M3 per-row-scaled rounding
-        // adds a few percent per table contribution; a mis-mapped table
-        // or a broken decode would still register ~27%, so 5% keeps the
-        // gate discriminating while tolerating quantization noise.
-        let fp8_tables: Vec<NgramTable> = ng
-            .tables
-            .iter()
-            .enumerate()
-            .map(|(index, t)| {
-                let NgramTable::Bf16(w) = t else {
-                    panic!("test built BF16 tables")
-                };
-                NgramTable::quantize_bf16(w, table_rows_n[index], td, g, stream).unwrap()
-            })
-            .collect();
-        let fp8_worst = worst_row(&run(&fp8_tables));
-        assert!(
-            fp8_worst < 0.05,
-            "FP8-quantized fused embedding diverges from golden: worst row \
-             Frobenius rel = {fp8_worst}"
-        );
-        println!("ngram GPU parity (FP8 tables): worst row Frobenius rel = {fp8_worst:.4}");
-    }
-
-    /// The NVMe-backed cache must be INVISIBLE: gathering through a bounded
-    /// row cache has to produce byte-identical output to the fully-resident
-    /// table it replaces, including under real pressure (fewer slots than
-    /// distinct rows, so rows are evicted and re-faulted mid-sequence).
-    ///
-    /// Uses the same compacted real-checkpoint data as the resident parity
-    /// test; the table file is written row-major so the cache's
-    /// `row_id -> byte offset` contract is exercised for real.
-    #[test]
-    #[ignore] // GPU + filesystem
-    fn cached_table_matches_resident_table() {
-        let dir = std::env::var("ATLAS_NGRAM_TEST_DATA")
-            .expect("set ATLAS_NGRAM_TEST_DATA (see make_gpu_testdata.py)");
-        let meta: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(format!("{dir}/meta.json")).unwrap())
-                .unwrap();
-        let gpu =
-            spark_runtime::cuda_backend::AtlasCudaBackend::new(0, &atlas_kernels::ptx_modules())
-                .expect("CUDA backend");
-        let g: &dyn GpuBackend = &gpu;
-        let stream = g.default_stream();
-
-        // Table 0's compacted rows: [n_rows, 256] BF16, row id == compact index.
-        let t0 = &meta["tables"].as_array().unwrap()[0];
-        let n_rows = t0["ids"].as_array().unwrap().len();
-        let dim = t0["dim"].as_u64().unwrap() as usize;
-        let bytes = std::fs::read(format!("{dir}/table0_rows.bin")).unwrap();
-        assert_eq!(bytes.len(), n_rows * dim * 2);
-        let row_stride = dim * 2; // 512 B — divides the 4 KiB O_DIRECT block
-
-        // Resident copy.
-        let resident = g.alloc(bytes.len()).unwrap();
-        g.copy_h2d_async(&bytes, resident, stream).unwrap();
-
-        // Backing file for the cached copy.
-        let tmp =
-            std::env::temp_dir().join(format!("atlas_ngram_cache_{}.bin", std::process::id()));
-        std::fs::write(&tmp, &bytes).unwrap();
-
-        // Deliberately fewer slots than rows -> eviction + re-fault.
-        let slots = (n_rows / 3).max(4);
-        let mut cache =
-            spark_storage::NgramRowCache::open(&tmp, None, n_rows as u64, row_stride, slots)
-                .expect("open cache");
-
-        // Walk every row twice (second pass exercises re-faulting), in
-        // BATCHES — a slot is pinned for the batch in flight, so the cache
-        // must hold at least one forward's worth of rows per table (the
-        // serving constraint: slots >= max tokens per forward).
-        let ids: Vec<u64> = (0..n_rows as u64).chain(0..n_rows as u64).collect();
-        let m = ids.len();
-        let batch = slots / 2;
-        let ids_dev = g.alloc(m * 4).unwrap();
-        let out_res = g.alloc(m * dim * 2).unwrap();
-        let out_cache = g.alloc(m * dim * 2).unwrap();
-
-        let embed_k = g.kernel("embed_from_argmax", "batched_embed").unwrap();
-        // Resident: gather by ROW id.
-        let idb: Vec<u8> = ids.iter().flat_map(|v| (*v as u32).to_le_bytes()).collect();
-        g.copy_h2d_async(&idb, ids_dev, stream).unwrap();
-        crate::layers::ops::batched_embed(
-            g, embed_k, ids_dev, resident, out_res, m as u32, dim as u32, stream,
-        )
-        .unwrap();
-
-        // Cached: per batch, resolve to SLOTS then gather from the pinned
-        // arena into that batch's slice of the output.
-        let table = DevicePtr(cache.table_dev_va().unwrap());
-        let mut slot_ids = Vec::new();
-        for (bi, chunk) in ids.chunks(batch).enumerate() {
-            cache.resolve(chunk, &mut slot_ids).unwrap();
-            assert_eq!(slot_ids.len(), chunk.len());
-            let sb: Vec<u8> = slot_ids.iter().flat_map(|v| v.to_le_bytes()).collect();
-            g.copy_h2d_async(&sb, ids_dev, stream).unwrap();
-            let dst = out_cache.offset(bi * batch * dim * 2);
-            crate::layers::ops::batched_embed(
-                g,
-                embed_k,
-                ids_dev,
-                table,
-                dst,
-                chunk.len() as u32,
-                dim as u32,
-                stream,
-            )
-            .unwrap();
-            // The gather is issued; the batch's pins can be released. Sync
-            // first so a later fault cannot overwrite a slot this gather is
-            // still reading (the serving path issues on one stream in order).
-            g.synchronize(stream).unwrap();
-            cache.end_batch();
-        }
-
-        let mut a = vec![0u8; m * dim * 2];
-        let mut b = vec![0u8; m * dim * 2];
-        g.copy_d2h(out_res, &mut a).unwrap();
-        g.copy_d2h(out_cache, &mut b).unwrap();
-        let _ = std::fs::remove_file(&tmp);
-        let (hits, misses, evictions) = cache.stats();
-        assert!(
-            evictions > 0,
-            "test must exercise eviction (slots={slots}, rows={n_rows}); \
-             hits={hits} misses={misses} evictions={evictions}"
-        );
-        assert_eq!(
-            a, b,
-            "cached gather diverged from the resident table \
-             (hits={hits} misses={misses} evictions={evictions})"
-        );
-        println!(
-            "ngram row cache: BYTE-IDENTICAL over {m} lookups with {slots}/{n_rows} slots \
-             (hits={hits} misses={misses} evictions={evictions})"
-        );
-    }
-
-    #[test]
-    fn decode_context_window_suffices() {
-        // Ids for the LAST token computed over the full sequence must equal
-        // ids computed over only the trailing (n-1)-token window + the new
-        // token — the NgramCache contract the serving path relies on.
-        let dims = NgramDims {
-            vocab_size: 997,
-            ratio: 5,
-            neighbor_num: 4,
-            split_num: 2,
-            eos_token_id: 2,
-            hidden_size: 48,
-        };
-        let seq: Vec<u32> = vec![901, 15, 371, 2, 88, 990, 41, 7, 640, 3, 55];
-        let full = ngram_ids(&dims, &seq);
-        let keep = dims.neighbor_num - 1;
-        let tail: Vec<u32> = seq[seq.len() - 1 - keep..].to_vec();
-        let win = ngram_ids(&dims, &tail);
-        for (index, ids) in full.iter().enumerate() {
-            assert_eq!(
-                ids.last(),
-                win[index].last(),
-                "window-decode diverges at table {index}"
-            );
-        }
-    }
-}
+mod tests;
