+++ crates/spark-model/src/model/mod.rs
@@ -30,6 +30,7 @@ pub(crate) mod impl_a1;
 pub(crate) mod impl_a1_init;
 pub(crate) mod impl_a2;
 pub(crate) mod impl_a3;
+mod impl_a3_embed;
 pub(crate) mod impl_b1;
 pub(crate) mod impl_b2;
 pub(crate) mod impl_b3;
